Move jackson mapper to separate object
